requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
google-generativeai
sentence-transformers
PyMuPDF
tqdm
python-dotenv